Added support for SLES 15

+++ Testscripts/Linux/perf_iperf3.sh
@@ -27,44 +27,79 @@ touch ./IPERF3Test.log
 
 InstallIPERF3()
 {
-		DISTRO=`grep -ihs "buntu\|Suse\|Fedora\|Debian\|CentOS\|Red Hat Enterprise Linux\|clear-linux-os" /etc/{issue,*release,*version} /usr/lib/os-release`
-		if [[ $DISTRO =~ "Ubuntu" ]];
-		then
-			
-			LogMsg "Detected Ubuntu"
-			ssh ${1} "until dpkg --force-all --configure -a; sleep 10; do echo 'Trying again...'; done"
-			ssh ${1} "apt-get update"
-			ssh ${1} "apt-get -y install iperf3 sysstat bc psmisc"
-			if [ $IPversion -eq 6 ]; then	
-				scp ConfigureUbuntu1604IPv6.sh ${1}:
-				ssh ${1} "chmod +x ConfigureUbuntu1604IPv6.sh"
-				ssh ${1} "./ConfigureUbuntu1604IPv6.sh"
-			fi
-		elif [[ $DISTRO =~ "Red Hat Enterprise Linux Server release 6" ]] || [[ $DISTRO =~ "CentOS Linux release 6" ]] || [[ $DISTRO =~ "CentOS release 6" ]];
-		then
-				LogMsg "Detected Redhat/CentOS 6.x"
-				ssh ${1} "rpm -ivh https://dl.fedoraproject.org/pub/epel/epel-release-latest-6.noarch.rpm"
-				ssh ${1} "yum -y --nogpgcheck install iperf3 sysstat bc psmisc"
-				ssh ${1} "iptables -F"				
-				
-		elif [[ $DISTRO =~ "Red Hat Enterprise Linux Server release 7" ]] || [[ $DISTRO =~ "CentOS Linux release 7" ]];
-		then
-				LogMsg "Detected Redhat/CentOS 7.x"
-				ssh ${1} "rpm -ivh https://dl.fedoraproject.org/pub/epel/epel-release-latest-7.noarch.rpm"
-				ssh ${1} "yum -y --nogpgcheck install iperf3 sysstat bc psmisc"
-				ssh ${1} "iptables -F"
-		elif [[ $DISTRO =~ "clear-linux-os" ]];
-		then
-				LogMsg "Detected Clear Linux OS. Installing required packages"
-				ssh ${1} "swupd bundle-add dev-utils-dev sysadmin-basic performance-tools os-testsuite-phoronix network-basic openssh-server dev-utils os-core os-core-dev"
-				ssh ${1} "iptables -F"
-								
-		else
-				LogMsg "Unknown Distro"
-				UpdateTestState "TestAborted"
-				UpdateSummary "Unknown Distro, test aborted"
-				return 1
+	DISTRO=`grep -ihs "ubuntu\|Suse\|Fedora\|Debian\|CentOS\|Red Hat Enterprise Linux\|clear-linux-os" /etc/{issue,*release,*version} /usr/lib/os-release`
+	if [[ $DISTRO =~ "Ubuntu" ]];
+	then	
+		LogMsg "Detected Ubuntu"
+		ssh ${1} "until dpkg --force-all --configure -a; sleep 10; do echo 'Trying again...'; done"
+		ssh ${1} "apt-get update"
+		ssh ${1} "apt-get -y install iperf3 sysstat bc psmisc"
+		if [ $IPversion -eq 6 ]; then	
+			scp ConfigureUbuntu1604IPv6.sh ${1}:
+			ssh ${1} "chmod +x ConfigureUbuntu1604IPv6.sh"
+			ssh ${1} "./ConfigureUbuntu1604IPv6.sh"
 		fi
+	elif [[ $DISTRO =~ "Red Hat Enterprise Linux Server release 6" ]] || [[ $DISTRO =~ "CentOS Linux release 6" ]] || [[ $DISTRO =~ "CentOS release 6" ]];
+	then
+		LogMsg "Detected Redhat/CentOS 6.x"
+		ssh ${1} "rpm -ivh https://dl.fedoraproject.org/pub/epel/epel-release-latest-6.noarch.rpm"
+		ssh ${1} "yum -y --nogpgcheck install iperf3 sysstat bc psmisc"
+		ssh ${1} "iptables -F"				
+	elif [[ $DISTRO =~ "Red Hat Enterprise Linux Server release 7" ]] || [[ $DISTRO =~ "CentOS Linux release 7" ]];
+	then
+		LogMsg "Detected Redhat/CentOS 7.x"
+		ssh ${1} "rpm -ivh https://dl.fedoraproject.org/pub/epel/epel-release-latest-7.noarch.rpm"
+		ssh ${1} "yum -y --nogpgcheck install iperf3 sysstat bc psmisc"
+		ssh ${1} "iptables -F"
+	elif [[ $DISTRO =~ "SUSE Linux Enterprise Server" ]];
+	then
+		LogMsg "Detected SLES"
+		if [[ $DISTRO =~ "SUSE Linux Enterprise Server 12" ]];
+		then
+			LogMsg "Detected SLES 12"
+			repositoryUrl="https://download.opensuse.org/repositories/network:utilities/SLE_12_SP3/network:utilities.repo"
+		elif [[ $DISTRO =~ "SUSE Linux Enterprise Server 15" ]];
+		then
+			LogMsg "Detected SLES 15"
+			repositoryUrl="https://download.opensuse.org/repositories/network:utilities/SLE_15/network:utilities.repo"
+		else
+			LogMsg "Error: Unknown SLES version"
+			UpdateTestState "TestAborted"
+			return 2
+		fi
+		ssh ${1} "zypper addrepo ${repositoryUrl}"
+		ssh ${1} "zypper --no-gpg-checks --non-interactive --gpg-auto-import-keys refresh"
+		ssh ${1} "zypper --no-gpg-checks --non-interactive --gpg-auto-import-keys install sysstat git bc make gcc psmisc"
+		ssh ${1} "which iperf3"
+		if [ $? -ne 0 ]; then
+			LogMsg "Info: iperf3 is not installed. So, Installing iperf3 using rpm"
+			iperfUrl="https://eosgpackages.blob.core.windows.net/testpackages/tools/iperf-sles-x86_64.rpm"
+			libIperfUrl="https://eosgpackages.blob.core.windows.net/testpackages/tools/libiperf0-sles-x86_64.rpm"
+			ssh ${1} "wget --no-check-certificate ${iperfUrl}"
+			ssh ${1} "wget --no-check-certificate ${libIperfUrl}"
+			LogMsg "rpm -ivh ${iperfUrl##*/} ${libIperfUrl##*/}"
+			ssh ${1} "rpm -ivh ${iperfUrl##*/} ${libIperfUrl##*/}"
+			ssh ${1} "which iperf3"
+			if [ $? -ne 0 ]; then
+				LogMsg "Error: Unable to install iperf3 from source/rpm"
+				UpdateTestState "TestAborted"
+				return 3
+			fi				
+		else
+			LogMsg "Info: Iperf3 installed from repository"
+		fi
+		ssh ${1} "iptables -F"	
+	elif [[ $DISTRO =~ "clear-linux-os" ]];
+	then
+		LogMsg "Detected Clear Linux OS. Installing required packages"
+		ssh ${1} "swupd bundle-add dev-utils-dev sysadmin-basic performance-tools os-testsuite-phoronix network-basic openssh-server dev-utils os-core os-core-dev"
+		ssh ${1} "iptables -F"					
+	else
+		LogMsg "Unknown Distro"
+		UpdateTestState "TestAborted"
+		UpdateSummary "Unknown Distro, test aborted"
+		return 1
+	fi
 }
 
 
@@ -177,10 +212,18 @@ fi
 
 LogMsg "Configuring client ${client}..."
 InstallIPERF3 ${client}
-
+if [ $? -ne 0 ]; then
+	LogMsg "Error: iperf installation failed in ${client}.."
+	UpdateTestState "TestAborted"
+	exit 1
+fi
 LogMsg "Configuring server ${server}..."
 InstallIPERF3 ${server}
-
+if [ $? -ne 0 ]; then
+	LogMsg "Error: iperf installation failed in ${server}.."
+	UpdateTestState "TestAborted"
+	exit 1
+fi
 ssh ${server} "rm -rf iperf-server-*"
 ssh ${client} "rm -rf iperf-client-*"
 ssh ${client} "rm -rf iperf-server-*"
